Move connect validation to IOHelper
